Update lobsterPARty
Removed player_name argument: The original function definition of fetch_data included an extra player_name argument that was not used. We updated the function definition to only have the url argument.

Installed missing packages: In the Jupyter Notebook environment, we installed the missing packages (requests, beautifulsoup4, matplotlib, and tensorflow) using the !pip install command.

Fixed the AttributeError: We updated the CSS selector for fetching the rows in the BeautifulSoup parsing, which resolved the AttributeError.

Fixed the IndexError: The 'data' field in the sequence_sets dictionary was wrapped in an extra list, causing the shape of the X array to be incorrect. We removed the extra brackets to fix the IndexError.

With these changes, the code now fetches the data from the given URL, prepares the data, trains an LSTM model on the data, and predicts the next number in the sequences for points, assists, rebounds, and 3-pointers made.